Implement Transcribing DNA into RNA task
