+++ andrey_petrov/src/com/marsofandrew/bioinformatic/tasks/CountingDNANucleotide.java
@@ -1,17 +1,15 @@
 package com.marsofandrew.bioinformatic.tasks;
 
+import com.marsofandrew.bioinformatic.common.Helpers;
+
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Scanner;
 
 public class CountingDNANucleotide {
 
     public static void main(String[] args) throws IOException {
-        try (Scanner scanner = new Scanner(System.in)) {
-            String dataSet = scanner.nextLine();
-            System.out.println(countDNANucleotides(dataSet));
-        }
+        Helpers.executeStringInput(CountingDNANucleotide::countDNANucleotides);
     }
 
     public static String countDNANucleotides(final String data) {
